% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{calculatePEHE}
\alias{calculatePEHE}
\title{Calculate Precision in Estimation Heterogeneous treatment Effects}
\usage{
calculatePEHE(data, predictedBenefit)
}
\arguments{
\item{data}{The simulated dataset. It must contain the following
columns: \emph{untreatedRiskLinearPredictor} contains
the linear predictor of risk if left untreated;
\emph{treatedRiskLinearPredictor} contains the linear
predictor of true risk if left untreated;
\emph{riskLinearPredictor} contains the predicted risk
linear predictor.}

\item{predictedBenefit}{A vector of predicted absolute benefits for \emph{data}.}
}
\description{
Estimate the unobservable squared loss of absolute treatment benefit
}
