% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createBaselineRiskSettings}
\alias{createBaselineRiskSettings}
\title{Create baseline risk settings}
\usage{
createBaselineRiskSettings(type = "binary", modelSettings)
}
\arguments{
\item{type}{The type of the outcome}

\item{modelSettings}{The settings required for the definition of the
baseline risk model}
}
\value{
A list of settings for the definition of the baseline risk model
}
\description{
Creates the settings required for defining the baseline risk model
}
