% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createBinomialDistributionSettings}
\alias{createBinomialDistributionSettings}
\title{Binomial distribution settings}
\usage{
createBinomialDistributionSettings(size = 1, prob)
}
\arguments{
\item{size}{Number of trials (zero or more).}

\item{prob}{Probability of success on each trial.}
}
\value{
A list with the settings for simulating from a binomial distribution.
}
\description{
Create the settings for sampling from a binomial distribution.
}
