% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createDatabaseSettings}
\alias{createDatabaseSettings}
\title{Create the database settings}
\usage{
createDatabaseSettings(
  numberOfObservations,
  numberOfCovariates,
  covariateDistributionSettings
)
}
\arguments{
\item{numberOfObservations}{The number of patients in the dataset}

\item{numberOfCovariates}{The number of covariates iun the dataset}

\item{covariateDistributionSettings}{A list containing the distribution settings for
each covariate in the dataset}
}
\value{
A list with the settings required to simulate a dataset
}
\description{
Holds the all the settings for simulating a dataset.
}
