% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createModelSettings}
\alias{createModelSettings}
\title{Create model settings}
\usage{
createModelSettings(
  type = "logistic",
  constant,
  modelMatrix,
  transformationSettings,
  coefficients
)
}
\arguments{
\item{type}{The type of the dependent variable}

\item{constant}{The constant term of the model}

\item{modelMatrix}{A logical matrix that instructs which covariates will be
used. Each row represents a term in the model constructed
multiplicatively based on the columns that represents
each one of the dataset's covariates.}

\item{transformationSettings}{A list of functions defining the transformation for each
term of the model}

\item{coefficients}{A vector with the coefficients of the model}
}
\value{
A list with the required settings for the definition of a model
}
\description{
Creates the settings required to define a model explaining the relationship
between covariates
}
