% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createMultivariateNormalDistributionSettings}
\alias{createMultivariateNormalDistributionSettings}
\title{Multivariate normal distribution settings}
\usage{
createMultivariateNormalDistributionSettings(mean, covariance)
}
\arguments{
\item{mean}{A vector with the means of the target normal distribution}

\item{covariance}{The covariance matrix of the target multivariate normal distribution}
}
\value{
a list with the settings for simulating from a multivariate normal distribution
}
\description{
create the settings for sampling from a multivariate normal distribution
}
