% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createNormalDistributionSettings}
\alias{createNormalDistributionSettings}
\title{Normal distribution settings}
\usage{
createNormalDistributionSettings(mean = 0, covariance = 1)
}
\arguments{
\item{mean}{the mean of the target normal distribution}

\item{covariance}{the covariance of the target normal distribution}
}
\value{
a list with the settings for simulating from a normal distribution
}
\description{
create the settings for sampling from a normal distribution
}
