% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createPropensitySettings}
\alias{createPropensitySettings}
\title{Create propensity score settings}
\usage{
createPropensitySettings(type = "binary", modelSettings)
}
\arguments{
\item{type}{The type of the treatment variable. Default is "binary"}

\item{modelSettings}{The settings required for the definition of the
propensity score model}
}
\value{
A list of settings fir the definition of the propensity score model
}
\description{
Creates the settings required for the definition of the model based on which
treatment is administered
}
