% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createTreatmentEffectSettings}
\alias{createTreatmentEffectSettings}
\title{Create treatment effect settings}
\usage{
createTreatmentEffectSettings(type = "lp", modelSettings)
}
\arguments{
\item{type}{The reference of treatment effect. If set to "lp", treatment
effect is modeled as a function of the linear predictor of
baseline risk}

\item{modelSettings}{The settings required for defining the model of treatment
effect}
}
\value{
A list of settings for the definition of the treatment effect model
}
\description{
Creates the settings required for modeling the effect of treatment
}
