% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{generateBaselineData}
\alias{generateBaselineData}
\title{Generate the baseline data}
\usage{
generateBaselineData(databaseSettings)
}
\arguments{
\item{databaseSettings}{The database settings defined using \code{\link[=createDatabaseSettings]{createDatabaseSettings()}}}
}
\value{
A dataframe with the observed covariates for each patient in the
simulated dataset
}
\description{
Generates the dataset of baseline covariates
}
