% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{generateLinearPredictor}
\alias{generateLinearPredictor}
\title{Generate linear predictor}
\usage{
generateLinearPredictor(data, modelSettings)
}
\arguments{
\item{data}{The dataframe based on which the model will be defined}

\item{modelSettings}{The model settings defined based on \code{\link[=createModelSettings]{createModelSettings()}}}
}
\value{
The linear predictor for each patient in the simulated dataset
}
\description{
Generates the linear predictor of a model based on model settings
}
