% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createNormalDistributionSettings}
\alias{createNormalDistributionSettings}
\title{Normal distribution settings}
\usage{
createNormalDistributionSettings(mean = 0, covariance = 1)
}
\arguments{
\item{mean}{The mean of the target normal distribution}

\item{covariance}{The covariance of the target normal distribution}
}
\value{
A list with the settings for simulating from a normal distribution
}
\description{
Create the settings for sampling from a normal distribution
}
